requirements.txt
# LLM & API
openai>=1.0.0
groq>=0.4.0

# Graph Database
neo4j>=5.0.0
# CyVer>=0.1.0  # Optional: requires Python 3.10+, fallback validators used if unavailable

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# NLP Metrics
rouge>=1.0.0
rouge_score>=0.1.2
nltk>=3.8.0
textdistance>=4.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
regex>=2023.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
